feat(runners): add configurable eviction strategy to idle config (#3375)
We do some on-instance caching so when we scale down we'd prefer to keep
the older instances around instead of the new ones (because they will
have a hotter cache). This adds a configurable setting to the idleConfig
to pick a sorting strategy. Never contributed to this repo, so please
tell me if I'm doing something wrong!

---------

Co-authored-by: Niek Palm <npalm@users.noreply.github.com>