docs: Document using labels with self-hosted runners (#3508)
Explanation of issue: There are many popular CI systems that exist out
in the world, not only GitHub Actions. It's probably common enough that
a developer comes to this project and needs to configure the Terraform
runners with GitHub Actions, while also having in their mind how other
systems operate, and those may be different. It's easy to imagine all
runner labels must match all job labels. Other systems are designed that
way. To quickly solve the problem this is a sort of cheatsheet.