feat: namespace EC2 tags (#3523)
## Problem
The runners are still using a some ec2 tags such as Type, Repo, Owner
for selecting runners controlled by the control plane. Since the tags
are not namespaced they can conflict existing tags used in an AWS
account. This PR is namespacing the tags like the other tags used by the
module with `ghr`.

## Verification

- [x] - deploy single runner (root) module. Check runners getting proper
started and terminated
- [x] - deploy multi runner module. Check runners getting proper started
and terminated