fix(webhook): Avoid jobs are accepted without labels (#3548)
## Descirption

When using the mult-runner with an exact match set to `true` worklfows
events with an empty array of labels matches the first matcher. This
cause not intented runners are created. This change avoids jobs with no
labels ar accepted.

The exactMatch is some bagage we ccarry from the past. I have aaded a
set of unit tests to show how the match is working. We should think
about removing the switch.