feat: add AMI house keeping lambda (#3570)
## Description
Quite common to build AMI's with packer. Cleaning up old images could be
a challenge. This housekeeping lambda can remove old AMI based on a few
criteria.

The AMI housekeepr is implemented as an extra terraform module with a
lambda that runs once a day. The housekeepr can be configured to exclude
images based on reference in SSM, and tags. Images. The module is
deisabled by default and can used as part of the main module,
multi-runner or stand alone.

This PR also updates outdated lambda dependencies. And updates terraform
workflows to 1.5.6 as main version.

## Tested

- [x] default example (with new housekeeper)
- [x] multi runner example

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>