feat: SSM housekeeper (#3577)
# Description
The runner module uses SSM to provide the JIT config or token to the
runner. In case the runner does not start healthy the SSM parameter is
not deleted. This PR adds a Lambda to remove by default SSM paramaters
in the token path that are older then a day.

The lambda will be deployed by default as part of the control plane and
manage the tokens in the path used by the scale-up runner function.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>
Co-authored-by: Scott Guymer <scott.guymer@philips.com>