feat: add failover to on-demand in case request is failing (#3409)
Adding the option to create on-demand instances in case spot is failing.

## Problem

This module either support spot or on-demand instances. When using spot
(default), creation can fail with several reasons. In case there is no
sufficient capacity on AWS it makes sens to request on-demand instances.
AWS does not support this kind of requests via the fleet API. This PR
addresses this problme and add the option (opt-in_ to create on-demand
instances in case of Insufficient capacity.

## Migrations

No migrations required

## Opt-in

Opt in by setting the variable`enable_runner_on_demand_failover`


## Verfication

Not easy to test the failover. But due to changes in multi-runner,
runner module as well lambda scale-up and pool. The following checks are
required

- [x] Ephemeral example with pool
- [x] Multi runner example

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>