feat: migrate webhook runner configuration to SSM (#3728)
This PR migrates the confugration for the webhook from environment
variables to SSM to avoid the maximum size of environment variables is
reached.

## Implementation

The webhook will read the configuration from SSM as json string. As long
the lambda is hot the configuration is cached to speed-up the lambda
time. In cases of configuration changes Lambda resources will be
re-created by Terraform to ensure no cached values are used.

fix: #3594

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>
Co-authored-by: Niek Palm <npalm@users.noreply.github.com>
Co-authored-by: Niek Palm <niek.palm@philips.com>