feat: Allow to disable runner max scaling check (#3849)
## Description
Allow to disable to maximum runner check when max runners is set to -1

## Background
The maximum check is calling the AWS API describeInstances. When having
a huge fleet of runners this call is expensive. Currently we have an
issue with this API, this option allows to avoid calling the
describeInstance API. Drawback is there will be nox safegouard for
scaling.

---------

Co-authored-by: forest-pr|bot <forest-pr[bot]@users.noreply.github.com>