fix(webhook): Don't set `${SQS_WORKFLOW_JOB_QUEUE}` to empty string (#3943)
We currently indicate that this feature is disabled by setting the env
var to an empty string. But instead we could omit it, to avoid polluting
the environment, and causing a warning to be logged.

This is a companion to #3942. That one stops the warning mentioned, this
one fixes the original cause by not setting the env var in the first
place. Both could be merged, ideally.