fix: remove fetch override for octokit and versions (#4042)
## Problem
Updating octkit auth app to 6.1.1. fails runtime. Not in our tests due
to most of the SDK layer is mocked.

Only information we catch is a generic Error.
```
 HttpError: response.json(...).catch is not a function
```

Via [issue](https://github.com/octokit/auth-app.js/issues/634)
@wolfy1339 provides feedback that the cause could be related to
overriding the fetch in octokit. In PR #3554 where tracing was
introduced also the fetch command was overriden, this due the fact the
node fetch was not working properly for tracing HTTP calls, which are in
our case the cals to GitHub. This problem seems now to be fixed in the
[AWS xray
sdk](https://github.com/aws/aws-xray-sdk-node/blob/master/CHANGELOG.md#370).

## Solution
The solution is quite simple. Just remove axios as fetch implementation
(override) and use the default provide by octokit.


## Testing
The problem was not detected in the unit tests. Reason is that the unit
test mocking the octokit SDK. In case of an ovrride we have doen. We
hsould have added a test using nock to mock the HTTP calls and not the
full SDK to catch problems like this in the unit test. By removing the
override, we can relay again on the test efforts done by octokit.

## References
- fix #3966