fix: align inlince policies (#4046)
## Description
Most of the lambdas are defining inline policies in the dedicated role
for the lambda. The naming of the inline policies is a bit chaotic. This
PR aligns the naming removes redundant parts of the naming like the the
lamba name or prefix. This is allready part of the role


fix #4045