refactor: split webhook accepting and queue dispatching logic (#4160)
## Description

This PR refacting the webhook lambda. It splits the webhook logic in a
part responsible to accept the event and a part to dispatch to the
runner queues. Secondly it refactor and removes duplicate logic in
tests.

This change should allow to split the lambda in two lambda and and use
the AWS EventBridge to accept events, and use the dispacher only for
delegating an event to a runner queueu.