feat: add spot termination handler (#4176)
## Description

This PR adds a lambda to to log and metric spot termination based on the
cloudtrail event `BidEvictedEvent`. The feature is experimental and
disabled by default.

## Future directions

The current implemenation only helps to make spot termination visible to
an admin team. For the future we want to enrich a runner with
information via tagging what job is active, This allows to let the
termination handler also inform the user by adding a job annotation once
a spot termination occurs.

## Migration

No migration is required. By default the watcher is disabled.

- logging for the watcher is changed
- resources will be recreated for notification warning watcher

---------

Co-authored-by: philips-labs-pr|bot <philips-labs-pr[bot]@users.noreply.github.com>
Co-authored-by: Stuart Pearson <1926002+stuartp44@users.noreply.github.com>