chore: Added debug for queued events function (#4312)
When working out a job is queued or not, we do it within a function.
However, when we do it within this function there is very little ability
to see what the response was when debugging is enabled. So this PR is
allowing us to see what was determined from the function.