feat(packer): Allow passing in an instance_profile to packer (#4278)
We ran into needing to pass a custom instance profile to packer in order
to get AWS IAM credentials for pre-pulling some ECR images.

I noticed what I think was a missing `instance_type` override variable
in the windows-core-2022 file. I'm not sure if that was an oversight, or
if it was omitted for a reason. It's been a long time since I've been
proficient in Windows, so I wasn't able to test that part of this very
well.

Thanks so much for this project, this has been immensely beneficial for
us.

---------

Co-authored-by: Niek Palm <npalm@users.noreply.github.com>